Docker compose management
Docker compose now managing (dev version only) of postgres and scraping
services.

Simply building and running the compose file will result in a new database
that is inflated with a web scrape of the last two days. This is amazing
for dev iteration speed and simplifying the setup process!

TODO: compose the web app as well? Then the entire platform would be
wrapped in one compose file and dev setup would be trivial.