Insert records with appropriate timezone
Before, we were inserting records and they were automatically treated as
UTC. This was incorrect, and ended up with our records being interpreted
as 6 hours ahead of when they really happened. This gave extra work to
the scjail server to undo. Let's just not do that moving forward (and I
already updated the database records)